WIP
Plan: make function dictionary global, environments have a reference
to it. Maybe it's functions tacked on, maybe it's a reference; whatever.

Copying and updating functions should be easier - the snapshots get
stored normally, but deepcopy of an environment keeps the same reference
to the runner. Restoring a snapshot means setting the delegate and
updating the function ast bodies based on newest editor changes.