fix error message disappearing on second onChange
Couldn't figure out why onChange was being triggered again,
but it should be able to be called whenever and figure out
if anything significant has changed. Clearing the error
shouldn't happen unless we're reparsing so we can get a new error.

Warning: there could be issues with runtime errors with this.
I can't find any.