WIP stdlib map env doesn't the correct runner set!
(this is causing the failing main test)
Somehow function envs need to be transformed & ingested by
the real runner when specifying the scopes of the runner.
Or there needs to be another way of loading these?
Loading the code each time instead of building them once
ahead of time perhaps?