fix lazycanvas in Safari
Removes __lookupgetter__ and __lookupSetting__ which are deprecated
as a method of detecting whether a property of a canvas context object
is a dynamic property.
Implements the saved callback which gets or sets a property more
simply by actually using the normal get/set syntax.

Not all of these properties are properly dealt with elsewhere in the
code, including not being serialized when saving the current context.