remember mouse clicks for 50ms
Since adding the frame limit (and before to some degre as well)
the is-mouse-down? style call (clicked) sometimes missed mousepresses
because the mouse would be pressed while the code was at a different
part of the program or not running and was no longer pressed by the
time the (clicked) call happened. Now a mousedown any time in the last
50ms will count towards (clicked) the first time it is called, but if
clicked is called again it'll return the real value of whether the mouse
is actually down or not, until another mouse click happens to reset the
50ms timer.

This doesn't feel like cheating too much because it's called "clicked",
not "button-down-now".